Update TabActivity to DealsFeedActivity
- Finished deals feed filter layout
- set DealsFeedActivity as default activity
