+++ app/src/main/java/com/amazingcoders_android/activities/base/NavDrawerActivity.java
@@ -13,8 +13,8 @@ import android.view.View;
 import android.widget.FrameLayout;
 
 import com.amazingcoders_android.R;
-import com.amazingcoders_android.activities.DealsFeedActivity;
 import com.amazingcoders_android.activities.FrontPageActivity;
+import com.amazingcoders_android.activities.DealsFeedActivity;
 import com.amazingcoders_android.activities.VenuesFeedActivity;
 import com.amazingcoders_android.dialogs.AlertDialogFactory;
 import com.amazingcoders_android.helpers.PreferencesStore;
@@ -82,10 +82,10 @@ public abstract class NavDrawerActivity extends BaseActivity implements Navigati
                     startActivity(new Intent(this, VenuesFeedActivity.class));
                     break;
                 case R.id.navigation_item_3:
+                    break;
                 case R.id.navigation_item_4:
                 case R.id.navigation_sub_item_1:
                 case R.id.navigation_sub_item_2:
-//                    startActivity(new Intent(this, FrontPageActivity.class));
                     mDrawerLayout.closeDrawers();
                     break;
                 case R.id.navigation_sub_item_3:
